Added restrict channel function
